try to make e2e teset run on github
