+++ scripts/run-tests-hnsw-e2e.sh
@@ -10,6 +10,18 @@ function _main()
 
     pushd "$(_get_path_to_monorepo)/iris-mpc-upgrade-hawk" || exit
     cargo test --release --test e2e_genesis -- --include-ignored
+
+    # create the keys with both AWSCURRENT and AWSPREVIOUS states
+    export AWS_ENDPOINT_URL="http://127.0.0.1:4566"
+    cargo run --release -p iris-mpc-bins --bin key-manager --  --region us-east-1 --node-id 0 --env dev rotate --public-key-bucket-name wf-dev-public-keys
+    cargo run --release -p iris-mpc-bins --bin key-manager --  --region us-east-1 --node-id 1 --env dev rotate --public-key-bucket-name wf-dev-public-keys
+    cargo run --release -p iris-mpc-bins --bin key-manager --  --region us-east-1 --node-id 2 --env dev rotate --public-key-bucket-name wf-dev-public-keys
+
+    cargo run --release -p iris-mpc-bins --bin key-manager --  --region us-east-1 --node-id 0 --env dev rotate --public-key-bucket-name wf-dev-public-keys
+    cargo run --release -p iris-mpc-bins --bin key-manager --  --region us-east-1 --node-id 1 --env dev rotate --public-key-bucket-name wf-dev-public-keys
+    cargo run --release -p iris-mpc-bins --bin key-manager --  --region us-east-1 --node-id 2 --env dev rotate --public-key-bucket-name wf-dev-public-keys
+
+    cargo test --release --test e2e_hawk -- --include-ignored
     popd || exit
 }
